CS_test now working. Need to fix output.
CS_test is now working. It is a void function but the CS_Array that it edits flips all unsatisfactory CSs to "NULL." Now need to make that info useable by as568.txt.. I'm thinking of using the string::as568 and searching for [1] which would be the cross section. if that CS is not on the CS_Array list, it is skipped.